backend-project/django/projects/Quiz_App/requirement.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==2.1.11
Django==3.2.6
django-cors-headers==3.7.0
django-nested-admin==3.3.3
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
pyparsing==3.0.6
python-decouple==3.4
python-monkey-business==1.0.0
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.7
